requirements.txt
numpy==1.23.5
opencv-contrib-python==4.7.0.72
PyYAML==6.0.2
scipy==1.15.3
